Fixes from Enrico
