+++ packages/patterns/src/components/patterns-menu-items/pattern-convert-button.js
@@ -1,7 +1,11 @@
 /**
  * WordPress dependencies
  */
-import { BlockSettingsMenuControls } from '@wordpress/block-editor';
+import {
+	BlockSettingsMenuControls,
+	store as blockEditorStore,
+} from '@wordpress/block-editor';
+import { serialize } from '@wordpress/blocks';
 import { useCallback, useState } from '@wordpress/element';
 import {
 	MenuItem,
@@ -17,11 +21,6 @@ import { __ } from '@wordpress/i18n';
 import { store as noticesStore } from '@wordpress/notices';
 
 /**
- * Internal dependencies
- */
-import { store } from '../../store';
-
-/**
  * Menu control to convert block(s) to patterns.
  *
  * @param {Object}   props           Component props.
@@ -32,8 +31,23 @@ export default function PatternConvertButton( { clientIds } ) {
 	const [ isModalOpen, setIsModalOpen ] = useState( false );
 	const [ title, setTitle ] = useState( '' );
 
-	const { __experimentalConvertBlocksToPattern: convertBlocksToPattern } =
-		useDispatch( store );
+	const convertBlocksToPattern =
+		( clientIDs, _title ) =>
+		async ( { registry } ) => {
+			const pattern = {
+				title: _title || __( 'Untitled Pattern' ),
+				content: serialize(
+					registry
+						.select( blockEditorStore )
+						.getBlocksByClientId( clientIDs )
+				),
+				status: 'publish',
+			};
+
+			await registry
+				.dispatch( 'core' )
+				.saveEntityRecord( 'postType', 'wp_block_pattern', pattern );
+		};
 
 	const { createSuccessNotice, createErrorNotice } =
 		useDispatch( noticesStore );
