+++ packages/patterns/src/index.js
@@ -1,2 +1,1 @@
-export { store } from './store';
 export * from './components';
