--- packages/patterns/src/store/index.js
@@ -1,22 +1,0 @@
-/**
- * WordPress dependencies
- */
-import { createReduxStore, register } from '@wordpress/data';
-
-/**
- * Internal dependencies
- */
-import * as actions from './actions';
-
-const STORE_NAME = 'core/patterns';
-
-/**
- * Store definition for the patterns namespace.
- *
- * @see https://github.com/WordPress/gutenberg/blob/HEAD/packages/data/README.md#createReduxStore
- *
- * @type {Object}
- */
-export const store = createReduxStore( STORE_NAME, { actions } );
-
-register( store );
